[refactoring] Fixed TODO's

+++ signature_builder/signature_builder_lib/hash_maker.cpp
@@ -8,11 +8,10 @@ namespace builder::crypto
         , m_digest()
     {
         Sha512Initialise(&m_context);
-        Sha512Update(&m_context, data->data(), data->size());
+        Sha512Update(&m_context, data->data(), static_cast<uint32_t>(data->size()));
         Sha512Finalise(&m_context, &m_digest);
     }
 
-    // TODO: Подумай, может все-таки unique?
     utils::BinaryBufferPtr HashMaker::getHash() const
     {
         return std::make_shared<utils::BinaryBuffer>(m_digest.bytes, m_digest.bytes + SHA512_DIGEST_SIZE);
